Refactor graph rag.
