+++ graphrag/general/index.py
@@ -13,117 +13,101 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-import logging
-import os
-from concurrent.futures import ThreadPoolExecutor
-import json
-from functools import reduce
+from functools import reduce, partial
 import networkx as nx
+
+from api import settings
 from api.db import LLMType
 from api.db.services.llm_service import LLMBundle
 from api.db.services.user_service import TenantService
 from graphrag.general.community_reports_extractor import CommunityReportsExtractor
-from graphrag.general.entity_resolution import EntityResolution
+from graphrag.entity_resolution import EntityResolution
+from graphrag.general.extractor import Extractor
 from graphrag.general.graph_extractor import GraphExtractor, DEFAULT_ENTITY_TYPES
-from graphrag.general.mind_map_extractor import MindMapExtractor
-from graphrag.utils import graph_merge
-from rag.nlp import rag_tokenizer
-from rag.utils import num_tokens_from_string
+from graphrag.utils import graph_merge, set_entity, get_relation, set_relation, get_entity, get_graph, set_graph, \
+    chunk_id
+from rag.nlp import rag_tokenizer, search
+
+class Dealer:
+    def __init__(self,
+                 extractor: Extractor,
+                 tenant_id: str,
+                 kb_id: str,
+                 chunks: list[tuple[str, str]],
+                 language,
+                 callback=None,
+                 entity_types=DEFAULT_ENTITY_TYPES
+                 ):
+        _, tenant = TenantService.get_by_id(tenant_id)
+        self.llm_bdl = LLMBundle(tenant_id, LLMType.CHAT, tenant.llm_id)
+        ext = extractor(self.llm_bdl, language=language,
+                             entity_types=entity_types,
+                             get_entity=partial(get_entity, tenant_id, kb_id),
+                             set_entity=partial(set_entity, tenant_id, kb_id),
+                             get_relation=partial(get_relation, tenant_id, kb_id),
+                             set_relation=partial(set_relation, tenant_id, kb_id)
+                             )
+        ents, rels = ext(chunks, callback)
+        self.graph = nx.Graph()
+        for en in ents:
+            self.graph.add_node(en["entity_name"],
+                           entity_type=en["entity_type"],
+                           description=en["description"])
+        for rel in rels:
+            self.graph.add_edge(
+                rel["src_id"],
+                rel["tgt_id"],
+                weight=rel["weight"],
+                description=rel["description"]
+            )
+
+        old_graph = get_graph(tenant_id, kb_id)
+        if old_graph is not None:
+            self.graph = reduce(graph_merge, [old_graph, self.graph])
 
 
-def build_knowledge_graph_chunks(tenant_id: str, chunks: list[str], callback, entity_types=DEFAULT_ENTITY_TYPES):
-    _, tenant = TenantService.get_by_id(tenant_id)
-    llm_bdl = LLMBundle(tenant_id, LLMType.CHAT, tenant.llm_id)
-    ext = GraphExtractor(llm_bdl)
-    left_token_count = llm_bdl.max_length - ext.prompt_token_count - 1024
-    left_token_count = max(llm_bdl.max_length * 0.6, left_token_count)
+class WithCommunity(Dealer):
+    def __init__(self,
+                 tenant_id: str,
+                 kb_id: str,
+                 chunks: list[tuple[str, str]],
+                 callback,
+                 language,
+                 entity_types=DEFAULT_ENTITY_TYPES
+                 ):
+        super().__init__(GraphExtractor, tenant_id, kb_id, chunks, callback, language, entity_types)
 
-    assert left_token_count > 0, f"The LLM context length({llm_bdl.max_length}) is smaller than prompt({ext.prompt_token_count})"
+        er = EntityResolution(self.llm_bdl)
+        graph = er(self.graph).output
 
-    BATCH_SIZE=4
-    texts, graphs = [], []
-    cnt = 0
-    max_workers = int(os.environ.get('GRAPH_EXTRACTOR_MAX_WORKERS', 50))
-    with ThreadPoolExecutor(max_workers=max_workers) as exe:
-        threads = []
-        for i in range(len(chunks)):
-            tkn_cnt = num_tokens_from_string(chunks[i])
-            if cnt+tkn_cnt >= left_token_count and texts:
-                for b in range(0, len(texts), BATCH_SIZE):
-                    threads.append(exe.submit(ext, ["\n".join(texts[b:b+BATCH_SIZE])], {"entity_types": entity_types}, callback))
-                texts = []
-                cnt = 0
-            texts.append(chunks[i])
-            cnt += tkn_cnt
-        if texts:
-            for b in range(0, len(texts), BATCH_SIZE):
-                threads.append(exe.submit(ext, ["\n".join(texts[b:b+BATCH_SIZE])], {"entity_types": entity_types}, callback))
+        set_graph(tenant_id, kb_id, graph)
 
-        callback(0.5, "Extracting entities.")
-        graphs = []
-        for i, _ in enumerate(threads):
-            graphs.append(_.result().output)
-            callback(0.5 + 0.1*i/len(threads), f"Entities extraction progress ... {i+1}/{len(threads)}")
+        for node_degree in graph.degree:
+            graph.nodes[str(node_degree[0])]["rank"] = int(node_degree[1])
 
-    graph = reduce(graph_merge, graphs) if graphs else nx.Graph()
-    er = EntityResolution(llm_bdl)
-    graph = er(graph).output
+        callback(0.6, "Extracting community reports.")
+        cr = CommunityReportsExtractor(self.llm_bdl)
+        cr = cr(graph, callback=callback)
+        self.community_structure = cr.structured_output
+        self.community_reports = cr.output
 
-    _chunks = chunks
-    chunks = []
-    for n, attr in graph.nodes(data=True):
-        if attr.get("rank", 0) == 0:
-            logging.debug(f"Ignore entity: {n}")
-            continue
-        chunk = {
-            "name_kwd": n,
-            "important_kwd": [n],
-            "title_tks": rag_tokenizer.tokenize(n),
-            "content_with_weight": json.dumps({"name": n, **attr}, ensure_ascii=False),
-            "content_ltks": rag_tokenizer.tokenize(attr["description"]),
-            "knowledge_graph_kwd": "entity",
-            "rank_int": attr["rank"],
-            "weight_int": attr["weight"]
-        }
-        chunk["content_sm_ltks"] = rag_tokenizer.fine_grained_tokenize(chunk["content_ltks"])
-        chunks.append(chunk)
+        for community, desc in zip(cr.structured_output, cr.output):
+            chunk = {
+                "title_tks": rag_tokenizer.tokenize(community["title"]),
+                "content_with_weight": desc,
+                "content_ltks": rag_tokenizer.tokenize(desc),
+                "knowledge_graph_kwd": "community_report",
+                "weight_flt": community["weight"],
+                "entities_kwd": community["entities"],
+                "important_kwd": community["entities"],
+                "kb_id": kb_id
+            }
+            chunk["content_sm_ltks"] = rag_tokenizer.fine_grained_tokenize(chunk["content_ltks"])
+            settings.docStoreConn.insert([{"id": chunk_id(chunk), **chunk}], search.index_name(tenant_id))
 
-    callback(0.6, "Extracting community reports.")
-    cr = CommunityReportsExtractor(llm_bdl)
-    cr = cr(graph, callback=callback)
-    for community, desc in zip(cr.structured_output, cr.output):
-        chunk = {
-            "title_tks": rag_tokenizer.tokenize(community["title"]),
-            "content_with_weight": desc,
-            "content_ltks": rag_tokenizer.tokenize(desc),
-            "knowledge_graph_kwd": "community_report",
-            "weight_flt": community["weight"],
-            "entities_kwd": community["entities"],
-            "important_kwd": community["entities"]
-        }
-        chunk["content_sm_ltks"] = rag_tokenizer.fine_grained_tokenize(chunk["content_ltks"])
-        chunks.append(chunk)
+        set_graph(tenant_id, kb_id, graph)
 
-    chunks.append(
-        {
-            "content_with_weight": json.dumps(nx.node_link_data(graph), ensure_ascii=False, indent=2),
-            "knowledge_graph_kwd": "graph"
-        })
 
-    callback(0.75, "Extracting mind graph.")
-    mindmap = MindMapExtractor(llm_bdl)
-    mg = mindmap(_chunks).output
-    if not len(mg.keys()):
-        return chunks
-
-    logging.debug(json.dumps(mg, ensure_ascii=False, indent=2))
-    chunks.append(
-        {
-            "content_with_weight": json.dumps(mg, ensure_ascii=False, indent=2),
-            "knowledge_graph_kwd": "mind_map"
-        })
-
-    return chunks
