+++ graphrag/light/graph_prompt.py
@@ -4,7 +4,6 @@ Reference:
  - [LightRag](https://github.com/HKUDS/LightRAG)
 """
 
-GRAPH_FIELD_SEP = "<SEP>"
 
 PROMPTS = {}
 
@@ -128,22 +127,6 @@ Output:
 ("content_keywords"{tuple_delimiter}"first contact, control, communication, cosmic significance"){completion_delimiter}
 #############################""",
 ]
-
-PROMPTS[
-    "summarize_entity_descriptions"
-] = """You are a helpful assistant responsible for generating a comprehensive summary of the data provided below.
-Given one or two entities, and a list of descriptions, all related to the same entity or group of entities.
-Please concatenate all of these into a single, comprehensive description. Make sure to include information collected from all the descriptions.
-If the provided descriptions are contradictory, please resolve the contradictions and provide a single, coherent summary.
-Make sure it is written in third person, and include the entity names so we the have full context.
-Use {language} as output language.
-
-#######
--Data-
-Entities: {entity_name}
-Description List: {description_list}
-#######
-"""
 
 PROMPTS[
     "entiti_continue_extraction"
