+++ graphrag/light/index.py
@@ -23,117 +23,9 @@ from api.db import LLMType
 from api.db.services.llm_service import LLMBundle
 from api.db.services.user_service import TenantService
 from graphrag.light.graph_extractor import GraphExtractor
-from graphrag.light.prompt import PROMPTS
-from graphrag.utils import graph_merge
+from graphrag.light.graph_prompt import PROMPTS
+from graphrag.utils import graph_merge, get_entity, set_entity, get_relation, set_relation, get_graph, set_graph
 from rag.nlp import rag_tokenizer, search
-
-
-def chunk_id(chunk):
-    return xxhash.xxh64((chunk["content_with_weight"] + chunk["kb_id"]).encode("utf-8")).hexdigest()
-
-
-def get_entity(tenant_id, kb_id, ent_name):
-    conds = {
-        "fields": ["content_with_weight"],
-        "size": 1,
-        "entity_kwd": [ent_name],
-        "knowledge_graph_kwd": ["entity"]
-    }
-    res = settings.retrievaler.search(conds, search.index_name(tenant_id), [kb_id])
-    for id in res.ids:
-        try:
-            return json.loads(res.field[id]["content_with_weight"])
-        except Exception:
-            continue
-
-
-def set_entity(tenant_id, kb_id, ent_name, meta):
-    chunk = {
-        "important_kwd": [ent_name],
-        "title_tks": rag_tokenizer.tokenize(ent_name),
-        "entity_kwd": [ent_name],
-        "knowledge_graph_kwd": "entity",
-        "entity_type_kwd": meta["entity_type"],
-        "content_with_weight": json.dumps(meta, ensure_ascii=False),
-        "content_ltks": rag_tokenizer.tokenize(meta["description"]),
-        "source_id": meta["source_id"],
-        "kb_id": kb_id
-    }
-    chunk["content_sm_ltks"] = rag_tokenizer.fine_grained_tokenize(chunk["content_ltks"])
-    res = settings.retrievaler.search({"entity_kwd": ent_name, "size": 1, "fields": []},
-                                      search.index_name(tenant_id), [kb_id])
-    if res.ids:
-        settings.docStoreConn.update({"entity_kwd": ent_name}, chunk, search.index_name(tenant_id), kb_id)
-    else:
-        settings.docStoreConn.insert([{"id": chunk_id(chunk), **chunk}], search.index_name(tenant_id))
-
-
-def get_relation(tenant_id, kb_id, from_ent_name, to_ent_name):
-    conds = {
-        "fields": ["content_with_weight"],
-        "size": 1,
-        "from_entity_kwd": [from_ent_name, to_ent_name],
-        "to_entity_kwd": [from_ent_name, to_ent_name],
-        "knowledge_graph_kwd": ["relation"]
-    }
-    res = settings.retrievaler.search(conds, search.index_name(tenant_id), [kb_id])
-    for id in res.ids:
-        try:
-            return json.loads(res.field[id]["content_with_weight"])
-        except Exception:
-            continue
-
-
-def set_relation(tenant_id, kb_id, from_ent_name, to_ent_name, meta):
-    chunk = {
-        "from_entity_kwd": [from_ent_name],
-        "to_entity_kwd": [to_ent_name],
-        "knowledge_graph_kwd": "relation",
-        "content_with_weight": json.dumps(meta, ensure_ascii=False),
-        "content_ltks": rag_tokenizer.tokenize(meta["description"]),
-        "important_kwd": meta["keywords"],
-        "source_id": meta["source_id"],
-        "weight_int": int(meta["weight"]),
-        "kb_id": kb_id
-    }
-    chunk["content_sm_ltks"] = rag_tokenizer.fine_grained_tokenize(chunk["content_ltks"])
-    res = settings.retrievaler.search({"from_entity_kwd": to_ent_name, "to_entity_kwd": to_ent_name, "size": 1, "fields": []},
-                                      search.index_name(tenant_id), [kb_id])
-    if res.ids:
-        settings.docStoreConn.update({"from_entity_kwd": to_ent_name, "to_entity_kwd": to_ent_name},
-                                 chunk,
-                                 search.index_name(tenant_id), kb_id)
-    else:
-        settings.docStoreConn.insert([{"id": chunk_id(chunk), **chunk}], search.index_name(tenant_id))
-
-
-def get_graph(tenant_id, kb_id):
-    conds = {
-        "fields": ["content_with_weight"],
-        "size": 1,
-        "knowledge_graph_kwd": ["graph"]
-    }
-    res = settings.retrievaler.search(conds, search.index_name(tenant_id), [kb_id])
-    for id in res.ids:
-        try:
-            return json_graph.node_link_graph(json.loads(res.field[id]["content_with_weight"]))
-        except Exception:
-            continue
-
-
-def set_graph(tenant_id, kb_id, graph):
-    chunk = {
-        "content_with_weight": json.dumps(nx.node_link_data(graph), ensure_ascii=False,
-                                          indent=2),
-        "knowledge_graph_kwd": "graph",
-        "kb_id": kb_id
-    }
-    res = settings.retrievaler.search({"knowledge_graph_kwd": "graph", "size": 1, "fields": []}, search.index_name(tenant_id), [kb_id])
-    if res.ids:
-        settings.docStoreConn.update({"knowledge_graph_kwd": "graph"}, chunk,
-                                     search.index_name(tenant_id), kb_id)
-    else:
-        settings.docStoreConn.insert([{"id": chunk_id(chunk), **chunk}], search.index_name(tenant_id))
 
 
 def build_knowledge_graph_with_chunks(tenant_id: str,
