+++ graphrag/light/smoke.py
@@ -16,14 +16,11 @@
 
 import argparse
 import json
-from functools import partial
-from api.db import LLMType
-from api.db.services.llm_service import LLMBundle
 from api import settings
 import networkx as nx
 from api.db.services.document_service import DocumentService
+from graphrag.general.index import Dealer
 from graphrag.light.graph_extractor import GraphExtractor
-from graphrag.light.index import get_entity, set_entity, get_relation, set_relation
 
 settings.init_settings()
 
@@ -41,26 +38,8 @@ if __name__ == "__main__":
     chunks = [d["content_with_weight"] for d in
               settings.retrievaler.chunk_list(args.doc_id, args.tenant_id, [kb_id], max_count=6,
                                               fields=["content_with_weight"])]
+    chunks = [("x", c) for c in chunks]
 
-    ext = GraphExtractor(LLMBundle(args.tenant_id, LLMType.CHAT), language="English",
-                         get_entity=partial(get_entity, args.tenant_id, kb_id),
-                         set_entity=partial(set_entity, args.tenant_id, kb_id),
-                         get_relation=partial(get_relation, args.tenant_id, kb_id),
-                         set_relation=partial(set_relation, args.tenant_id, kb_id)
-                         )
-    ents, rels = ext([("x", c) for c in chunks])
+    dealer = Dealer(GraphExtractor, args.tenant_id, kb_id, chunks, "English")
 
-    graph = nx.Graph()
-    for en in ents:
-        graph.add_node(en["entity_name"],
-                       entity_type=en["entity_type"],
-                       description=en["description"])
-    for rel in rels:
-        graph.add_edge(
-            rel["src_id"],
-            rel["tgt_id"],
-            weight=rel["weight"],
-            description=rel["description"]
-        )
-
-    print(json.dumps(nx.node_link_data(graph), ensure_ascii=False, indent=2))
+    print(json.dumps(nx.node_link_data(dealer.graph), ensure_ascii=False, indent=2))
